Refactor: InfluxDB 버전변경

+++ src/main/java/com/nhnacademy/sensordata/entity/humidity/Humidity.java
@@ -24,5 +24,5 @@ public class Humidity {
     @Column(name = "topic")
     private String topic;
     @Column(name = "value")
-    private Double value;
+    private Float value;
 }
